+++ src/SFA.DAS.PayeCreation.Tests/Project/Tests/StepDefinitions/PayeCreationSteps.cs
@@ -1,5 +1,4 @@
 ﻿using TechTalk.SpecFlow;
-using TechTalk.SpecFlow.Assist;
 
 namespace SFA.DAS.PayeCreation.Project.Tests.StepDefinitions
 {
@@ -29,20 +28,5 @@ namespace SFA.DAS.PayeCreation.Project.Tests.StepDefinitions
 
         [Given(@"I add levy declarations")]
         public void GivenIAddLevyDeclarations() => _payeCreationStepshelper.AddLevyPaye();
-    }
-
-    [Binding]
-    public class SpecificPayeCreationSteps
-    {
-        private readonly ScenarioContext _context;
-        public SpecificPayeCreationSteps(ScenarioContext context) => _context = context;
-
-        [Given(@"I add levy declarations")]
-        public void GivenIAddLevyDeclarations(Table table) => GetPayeCreationStepshelper(table).AddLevyPaye();
-        
-        [Given(@"I add non levy declarations")]
-        public void GivenIAddNonLevyDeclarations(Table table) => GetPayeCreationStepshelper(table).AddNonLevyPaye();
-
-        private PayeCreationStepshelper GetPayeCreationStepshelper(Table table) => new PayeCreationStepshelper(_context, table.CreateInstance<PayeDetails>());
     }
 }
